Rewrote AddPhotoStep to tsx format and fix the sonar issues
